requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
bcrypt
pyjwt
openai
pdfplumber
python-docx
python-multipart
JinJa2
reportlab
